Refactor Error Handling Class
